refactor: 타입 수정

+++ src/main/java/com/example/budongbudong/domain/auctionwinner/repository/AuctionWinnerRepository.java
@@ -3,7 +3,7 @@ package com.example.budongbudong.domain.auctionwinner.repository;
 import com.example.budongbudong.domain.auctionwinner.entity.AuctionWinner;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuctionWinnerRepository extends JpaRepository<AuctionWinner,Integer> {
+public interface AuctionWinnerRepository extends JpaRepository<AuctionWinner, Long> {
 
     boolean existsByAuctionId(Long id);
 }
